api semi final

+++ database/migrations/2025_12_16_064206_create_aturan_fuzzies_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('aturan_fuzzies', function (Blueprint $table) {
             $table->id();
             $table->string('nama_aturan');
-            $table->unsignedBigInteger('nilai');
+            $table->float('nilai');
             $table->timestamps();
         });
     }
